fix event create

+++ domain/user.go
@@ -3,7 +3,7 @@ package domain
 import "time"
 
 type User struct {
-    ID        int64     `gorm:"primaryKey;autoIncrement" json:"id"`
+    ID        uint     `gorm:"primaryKey;autoIncrement" json:"id"`
     Username  string    `gorm:"size:100;not null" json:"username"`
     Fullname  string    `gorm:"size:100;not null" json:"fullname"`
     Profile   string    `gorm:"size:255" json:"profile"`
